+++ src/main/java/com/greedy/mokkoji/api/recruitment/dto/response/createRecruitment/CreateRecruitmentResponse.java
@@ -7,9 +7,9 @@ import java.util.List;
 @Builder
 public record CreateRecruitmentResponse(
         Long id,
-        List<String> imageUrls
+        List<String> uploadImageUrls
 ) {
-    public static CreateRecruitmentResponse of(Long id, List<String> imageUrls) {
-        return new CreateRecruitmentResponse(id, imageUrls);
+    public static CreateRecruitmentResponse of(Long id, List<String> uploadImageUrls) {
+        return new CreateRecruitmentResponse(id, uploadImageUrls);
     }
 }
