+++ src/main/java/com/greedy/mokkoji/api/recruitment/dto/response/updateRecruitment/UpdateRecruitmentResponse.java
@@ -4,9 +4,11 @@ import java.util.List;
 
 public record UpdateRecruitmentResponse(
         Long recruitmentId,
-        List<String> imageUrls
+        List<String> deleteImageUrls,
+        List<String> uploadImageUrls
 ) {
-    public static UpdateRecruitmentResponse of(Long recruitmentId, List<String> imageUrls) {
-        return new UpdateRecruitmentResponse(recruitmentId, imageUrls);
+    public static UpdateRecruitmentResponse of(Long recruitmentId, List<String> deleteImageUrls, List<String> uploadImageUrls) {
+        return new UpdateRecruitmentResponse(recruitmentId, deleteImageUrls, uploadImageUrls);
     }
 }
+
